Move runMLXWinogradTests into dedicated mlxtests.cpp
Extracts the MLX Winograd test entry point out of the production
backend translation unit. runMLXBatchNormFP16Test /
runMLXConvLayerFP16WinogradTest stay in mlxbackend.cpp (they use
file-local classes) and are called via forward declarations.
